修改：     .vscode/settings.json 修改：     ArduCopter/Copter.cpp 修改：     ArduCopter/Copter.h 新文件：   ArduCopter/dumpcore.sh_arducopter.139269.out 新文件：   ArduCopter/dumpcore.sh_arducopter.149252.out 新文件：   ArduCopter/dumpcore.sh_arducopter.74516.out 新文件：   ArduCopter/dumpstack.sh_arducopter.139269.out 新文件：   ArduCopter/dumpstack.sh_arducopter.149252.out 新文件：   ArduCopter/dumpstack.sh_arducopter.74516.out 新文件：   ArduCopter/mav_0_1.parm 新文件：   ardupilot.und/id.txt 新文件：   ardupilot.und/settings.xml 新文件：   ardupilot1.und/id.txt 新文件：   ardupilot1.und/info.json 新文件：   ardupilot1.und/settings.xml 新文件：   ardupilot2.und/id.txt 新文件：   ardupilot2.und/settings.xml 修改：     libraries/AC_AttitudeControl/AC_AttitudeControl_Multi.cpp 修改：     libraries/AP_Scheduler/AP_Scheduler.cpp 修改：     libraries/AP_Scheduler/AP_Scheduler.h 修改：     libraries/AP_Vehicle/AP_Vehicle.cpp 修改：     libraries/GCS_MAVLink/GCS_Common.cpp 新文件：   libraries/fuzz/fuzz.h

+++ libraries/GCS_MAVLink/GCS_Common.cpp
@@ -5089,7 +5089,7 @@ void GCS_MAVLINK::send_sys_status()
         control_sensors_present,
         control_sensors_enabled,
         control_sensors_health,
-        static_cast<uint16_t>(AP::scheduler().load_average() * 1000),
+        static_cast<uint64_t>(AP::scheduler().load_average() * 1000),
 #if !defined(HAL_BUILD_AP_PERIPH) || defined(HAL_PERIPH_ENABLE_BATTERY)
         battery.gcs_voltage() * 1000,  // mV
         battery_current,        // in 10mA units
